{"types":"C","runs":5}

+++ Problem16.java
@@ -3,9 +3,9 @@ import acm.program.ConsoleProgram;
 public class Problem16 extends ConsoleProgram{
 	
 	public void run() {
-		int a = readInt("Enter a: ");
-		int b = readInt("Enter b: ");
-		int c = readInt("Enter c: ");
+		double a = readInt("Enter a: ");
+		double b = readInt("Enter b: ");
+		double c = readInt("Enter c: ");
 		
 		println(1 / (1/a + 1/b + 1/c));
